backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
PyMySQL==1.1.1
SQLAlchemy==2.0.40
alembic==1.15.2
python-dotenv==1.1.0
passlib==1.7.4
pydantic==2.11.4
PyJWT==2.10.1
python-multipart==0.0.20
cryptography==44.0.3
requests==2.32.3
beautifulsoup4==4.13.4
httpx==0.28.1
APScheduler==3.11.0
user-agents==2.2.0
